+++ shardingsphere-governance/shardingsphere-governance-context/src/main/java/org/apache/shardingsphere/governance/context/authority/listener/factory/PrivilegeNodeChangedListenerBuilder.java
@@ -15,29 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.registry.listener.factory.impl;
+package org.apache.shardingsphere.governance.context.authority.listener.factory;
 
+import org.apache.shardingsphere.governance.context.authority.listener.PrivilegeNodeChangedListener;
 import org.apache.shardingsphere.governance.core.registry.listener.GovernanceListener;
-import org.apache.shardingsphere.governance.core.registry.listener.factory.GovernanceListenerFactory;
-import org.apache.shardingsphere.governance.core.registry.listener.impl.MetaDataChangedListener;
+import org.apache.shardingsphere.governance.core.registry.listener.builder.GovernanceListenerBuilder;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
 
-import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 
 /**
- * Meta data changed listener factory.
+ * Privilege changed listener builder.
  */
-public final class MetaDataChangedListenerFactory implements GovernanceListenerFactory {
+public final class PrivilegeNodeChangedListenerBuilder implements GovernanceListenerBuilder {
     
     @Override
     public GovernanceListener create(final RegistryCenterRepository repository, final Collection<String> schemaNames) {
-        return new MetaDataChangedListener(repository);
+        return new PrivilegeNodeChangedListener(repository);
     }
     
     @Override
     public Collection<Type> getWatchTypes() {
-        return Arrays.asList(Type.UPDATED, Type.ADDED, Type.DELETED);
+        return Collections.singleton(Type.UPDATED);
     }
 }
