+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/listener/builder/impl/DataSourceStateChangedListenerBuilder.java
@@ -15,29 +15,29 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.context.authority.listener.factory;
+package org.apache.shardingsphere.governance.core.registry.listener.builder.impl;
 
-import org.apache.shardingsphere.governance.context.authority.listener.PrivilegeNodeChangedListener;
 import org.apache.shardingsphere.governance.core.registry.listener.GovernanceListener;
-import org.apache.shardingsphere.governance.core.registry.listener.factory.GovernanceListenerFactory;
+import org.apache.shardingsphere.governance.core.registry.listener.builder.GovernanceListenerBuilder;
+import org.apache.shardingsphere.governance.core.registry.listener.impl.DataSourceStateChangedListener;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
 
+import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 
 /**
- * Privilege changed listener factory.
+ * Data source state changed listener builder.
  */
-public final class PrivilegeNodeChangedListenerFactory implements GovernanceListenerFactory {
+public final class DataSourceStateChangedListenerBuilder implements GovernanceListenerBuilder {
     
     @Override
     public GovernanceListener create(final RegistryCenterRepository repository, final Collection<String> schemaNames) {
-        return new PrivilegeNodeChangedListener(repository);
+        return new DataSourceStateChangedListener(repository, schemaNames);
     }
     
     @Override
     public Collection<Type> getWatchTypes() {
-        return Collections.singleton(Type.UPDATED);
+        return Arrays.asList(Type.ADDED, Type.UPDATED, Type.DELETED);
     }
 }
