+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/listener/builder/impl/GlobalRuleChangedListenerBuilder.java
@@ -15,32 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.registry.listener.factory;
+package org.apache.shardingsphere.governance.core.registry.listener.builder.impl;
 
 import org.apache.shardingsphere.governance.core.registry.listener.GovernanceListener;
-import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
+import org.apache.shardingsphere.governance.core.registry.listener.builder.GovernanceListenerBuilder;
+import org.apache.shardingsphere.governance.core.registry.listener.impl.GlobalRuleChangedListener;
 import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
+import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 
 import java.util.Collection;
+import java.util.Collections;
 
 /**
- * Governance listener factory.
+ *  Global rule changed listener builder.
  */
-public interface GovernanceListenerFactory {
+public final class GlobalRuleChangedListenerBuilder implements GovernanceListenerBuilder {
+    @Override
+    public GovernanceListener create(final RegistryCenterRepository repository, final Collection<String> schemaNames) {
+        return new GlobalRuleChangedListener(repository);
+    }
     
-    /**
-     * Create governance listener.
-     * 
-     * @param repository registry center repository
-     * @param schemaNames schema names
-     * @return governance listener
-     */
-    GovernanceListener create(RegistryCenterRepository repository, Collection<String> schemaNames);
-    
-    /**
-     * Get watch types.
-     * 
-     * @return watch types
-     */
-    Collection<Type> getWatchTypes();
+    @Override
+    public Collection<Type> getWatchTypes() {
+        return Collections.singleton(Type.UPDATED);
+    }
 }
